% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genXdata.R
\name{genXdata}
\alias{genXdata}
\title{genXdata}
\usage{
genXdata(
  n,
  nvar = 1,
  mu = rep(0, nvar),
  Sigma = diag(length(mu)),
  varnames = paste("x", 1:length(mu), sep = ""),
  roundto = NULL
)
}
\description{
genXdata
}
