% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urnsamples.data.frame.R
\name{urnsamples.data.frame}
\alias{urnsamples.data.frame}
\title{urnsamples.data.frame}
\usage{
\method{urnsamples}{data.frame}(x, size, replace = FALSE, ordered = FALSE, ...)
}
\description{
urnsamples.data.frame
}
